sketched in partial refactor to support creating arbitrary content manager groups
