+++ src/cli/handlers/sso/detail.mjs
@@ -4,9 +4,8 @@ import { getCredentials } from '../../../lib/shared/authentication-lib'
 const handler = async ({ db, globalOptions }) => {
   const dbDetails = db.sso.details
   if (dbDetails.identityStoreID !== undefined) {
-    return { success: true, data: dbDetails }
-  }
-  else { // we'll try and and return the details
+    return { success : true, data : dbDetails }
+  } else { // we'll try and and return the details
     const credentials = getCredentials(globalOptions)
 
     const findIdentityStoreResults = await findIdentityStoreStaged({ credentials })
@@ -14,7 +13,7 @@ const handler = async ({ db, globalOptions }) => {
 
     db.sso.details = findIdentityStoreResults
 
-    return { success: true, data: findIdentityStoreResults }
+    return { success : true, data : findIdentityStoreResults }
   }
 }
