+++ src/cli/lib/handle-get-iam-policy.mjs
@@ -1,4 +1,5 @@
 import commandLineArgs from 'command-line-args'
+import { Questioner } from 'question-and-answer'
 
 import { cliSpec } from '../constants'
 import { generateIAMPolicy } from '../../lib/shared/generate-iam-policy'
@@ -19,9 +20,34 @@ const instructions =
 const handleGetIAMPolicy = async ({ argv, db, globalOptions }) => {
   const getIAMPolicyOptionsSpec = getOptionsSpec({ cliSpec, name : 'get-iam-policy' })
   const getIAMPolicyOptions = commandLineArgs(getIAMPolicyOptionsSpec, { argv })
-  const withInstructions = getIAMPolicyOptions['with-instructions']
+  let {
+    'with-instructions' : withInstructions,
+    'group-name' : groupName
+  } = getIAMPolicyOptions
 
-  progressLogger.write(JSON.stringify(await generateIAMPolicy({ db, globalOptions }), null, '  ') + '\n')
+  const groupOptions = Object.keys(db.sso.groups) || []
+  if (groupOptions.length === 0) {
+    throw new Error('No groups defined in database. Try:\n\ncloudsite import')
+  }
+
+  if (groupName === undefined) {
+    const interrogationBundle = {
+      actions : [
+        {
+          prompt    : 'Select the group to get the IAM policy for:',
+          parameter : 'group-name',
+          options   : groupOptions
+        }
+      ]
+    }
+    const questioner = new Questioner({ interrogationBundle, output : progressLogger })
+    await questioner.question()
+    groupName = questioner.get('group-name')
+  } else if (!groupOptions.includes(groupName)) {
+    throw new Error(`Unknown group '${groupName}'. Is the local database out of date? Try:\n\ncloudsite import`)
+  }
+
+  progressLogger.write(JSON.stringify(await generateIAMPolicy({ db, globalOptions, groupName }), null, '  ') + '\n')
 
   if (withInstructions === true) {
     // As of 2024-03-29, smartIndent does not work with numbered lists, but it should in the future.
